changes to ensure Checkout can run without turbo.
