+++ android/build.gradle
@@ -2,13 +2,17 @@ group 'com.razorpay.razorpay_flutter'
 version '1.1.4'
 
 buildscript {
+    ext {
+        kotlin_version = '2.0.0'
+    }
     repositories {
         mavenCentral()
         google()
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath 'com.android.tools.build:gradle:7.1.3'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
 
@@ -20,12 +24,15 @@ rootProject.allprojects {
 }
 
 apply plugin: 'com.android.library'
+apply plugin: 'org.jetbrains.kotlin.android'
 
 android {
-    compileSdkVersion 31
+    compileSdkVersion 34
+    namespace 'com.razorpay.razorpay_flutter'
 
     defaultConfig {
         minSdkVersion 21
+        targetSdkVersion 34
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
     lintOptions {
@@ -35,5 +42,9 @@ android {
         implementation fileTree(dir: 'libs', include: ['*.jar'])
         api 'com.razorpay:checkout:1.6.+'
         implementation 'com.google.code.gson:gson:2.8.8'
+        implementation fileTree(dir: 'upi-turbo/build/outputs/', include: ['*-debug.aar'])
+    }
+    kotlinOptions {
+        jvmTarget = '17'
     }
 }
