+++ android/settings.gradle
@@ -1,1 +1,2 @@
 rootProject.name = 'razorpay_flutter'
+include ':upi-turbo'
