+++ example/android/.gradle/config.properties
@@ -1,0 +1,2 @@
+#Wed Jun 12 00:04:06 IST 2024
+java.home=/Applications/Android Studio.app/Contents/jbr/Contents/Home
