+++ example/android/app/build.gradle
@@ -79,17 +79,8 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.5.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    // Turbo sdk
-    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
-    implementation 'com.squareup.okhttp3:logging-interceptor:4.10.0'
-    implementation 'com.google.code.gson:gson:+'
-    implementation 'com.github.bumptech.glide:glide:4.15.1'
-    implementation  'io.sentry:sentry:6.21.0'
-    implementation 'com.segment.analytics.kotlin:android:1.10.2'
+    implementation fileTree(dir: '/Users/vivek.shindhe/Razorpay/turbo/android-turbo-sample-app/app/libs', include: ['*.aar'])
 
     kapt 'com.android.databinding:compiler:3.1.4'
 
-    implementation 'com.google.android.gms:play-services-auth:20.2.0'
-    implementation 'com.google.android.gms:play-services-auth-api-phone:18.0.1'
 }
