+++ example/android/local.properties
@@ -1,5 +1,5 @@
-sdk.dir=/Users/shibu.prasad/Library/Android/sdk
-flutter.sdk=/Users/shibu.prasad/Developer/flutter
+sdk.dir=/Users/vivek.shindhe/Library/Android/sdk
+flutter.sdk=/Users/vivek.shindhe/fvm/versions/3.3.4
 flutter.buildMode=debug
 flutter.versionName=1.0.0
 flutter.versionCode=1
